backend : Test cases

+++ springapp/src/main/java/com/examly/springapp/controller/PlantController.java
@@ -1,6 +1,7 @@
 package com.examly.springapp.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,20 +14,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.examly.springapp.dto.PlantData;
+import com.examly.springapp.dto.PlantRequest;
 import com.examly.springapp.model.Plant;
-import com.examly.springapp.service.PlantServiceImpl;
+import com.examly.springapp.repository.PlantRepo;
+import com.examly.springapp.service.PlantService;
 
 @RestController
 @RequestMapping("/api")
 public class PlantController {
-    private final PlantServiceImpl plantService;
+    private final PlantService plantService;
 
-    public PlantController(PlantServiceImpl plantService){
+    public PlantController(PlantService plantService){
         this.plantService = plantService;
     }
+
     @PostMapping("/plants")
-    public ResponseEntity<Plant> addPlant(@RequestBody PlantData plantData){
+    public ResponseEntity<Plant> addPlant(@RequestBody PlantRequest plantData){
         return new ResponseEntity<>(plantService.addPlant(plantData), HttpStatus.OK);
     }
 
@@ -42,12 +45,22 @@ public class PlantController {
 
     @DeleteMapping("/plants/{id}")
     public ResponseEntity<String> deletePlant(@PathVariable long id){
+        Optional<Plant> plant = plantService.getPlantById(id);
+        if(plant.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         plantService.deletePlant(id);
-        return new ResponseEntity<>("Deleted",HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @GetMapping("/plants/plan")
-    public ResponseEntity<List<Plant>> generateCarePlan(@RequestParam String method){
-        return new ResponseEntity<>(plantService.generateCarePlan(method),HttpStatus.OK);
+    public ResponseEntity<List<Plant>> generateCarePlan(@RequestParam(required = false) String method){
+        
+        List<Plant> carePlan = plantService.generateCarePlan(method);
+        if(carePlan == null){
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(carePlan,HttpStatus.OK);
+       
     }
 }
